Add model access subresource tests - (Task #952)
Extracted from ModelAccessResourceTest
- Add abstract test class
- Add new subresources to DynamicRepositoryFilterBinding

---
Task #952: Refactor ModelAccessResource
