+++ de.dlr.sc.virsat.server.test/src/de/dlr/sc/virsat/server/test/AllTests.java
@@ -26,6 +26,9 @@ import de.dlr.sc.virsat.server.repository.ServerRepositoryTest;
 import de.dlr.sc.virsat.server.resources.DocumentationResourceTest;
 import de.dlr.sc.virsat.server.resources.ModelAccessResourceTest;
 import de.dlr.sc.virsat.server.resources.ProjectManagementResourceTest;
+import de.dlr.sc.virsat.server.resources.modelaccess.CategoryAssignmentResourceTest;
+import de.dlr.sc.virsat.server.resources.modelaccess.PropertyResourceTest;
+import de.dlr.sc.virsat.server.resources.modelaccess.StructuralElementInstanceResourceTest;
 import junit.framework.JUnit4TestAdapter;
 
 /**
@@ -43,6 +46,9 @@ import junit.framework.JUnit4TestAdapter;
 	ProjectManagementResourceTest.class,
 	LoginServiceFactoryTest.class,
 	ModelAccessResourceTest.class,
+	PropertyResourceTest.class,
+	CategoryAssignmentResourceTest.class,
+	StructuralElementInstanceResourceTest.class,
 	TransactionalJsonProviderTest.class,
 	RepositoryUtilityTest.class,
 	DocumentationResourceTest.class,
