+++ de.dlr.sc.virsat.server/src/de/dlr/sc/virsat/server/auth/filter/DynamicRepositoryFilterBinding.java
@@ -13,7 +13,7 @@ import javax.ws.rs.container.DynamicFeature;
 import javax.ws.rs.container.ResourceInfo;
 import javax.ws.rs.core.FeatureContext;
 
-import de.dlr.sc.virsat.server.resources.ModelAccessResource.RepoModelAccessResource;
+import de.dlr.sc.virsat.server.resources.ModelAccessResource;
 
 /**
  * Binds the RepositoryFilter to specified classes and functions,
@@ -23,7 +23,7 @@ public class DynamicRepositoryFilterBinding implements DynamicFeature {
 
 	@Override
 	public void configure(ResourceInfo resourceInfo, FeatureContext context) {
-		if (RepoModelAccessResource.class.equals(resourceInfo.getResourceClass())) {
+		if (ModelAccessResource.RESOURCE_CLASSES.contains(resourceInfo.getResourceClass())) {
 			context.register(RepositoryFilter.class);
 		}
 	}
